# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clippzi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts (with optional filters)
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: string
            enum: [video, image, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/trending:
    get:
      operationId: getTrendingPosts
      tags: [posts]
      summary: Get trending posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like or unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  # ── FEED ──
  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Get personalized for-you feed
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /feed/following:
    get:
      operationId: getFollowingFeed
      tags: [feed]
      summary: Get feed from followed users
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
...
  /comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ──
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isLive: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        postCount: { type: integer }
        totalViews: { type: integer }
        role: { type: string, enum: [user, streamer, admin] }
        createdAt: { type: string }
...
    Post:
      type: object
      required: [id, userId, type, createdAt, likeCount, commentCount, viewCount]
      properties:
        id: { type: integer }
        userId: { type: integer }
        user: { $ref: "#/components/schemas/User" }
        type: { type: string, enum: [video, image] }
        title: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        mediaUrl: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        musicTitle: { type: ["string", "null"] }
        musicArtist: { type: ["string", "null"] }
        musicUrl: { type: ["string", "null"] }
        duration: { type: ["integer", "null"] }
        likeCount: { type: integer }
        commentCount: { type: integer }
        viewCount: { type: integer }
        shareCount: { type: integer }
        tags:
          type: array
          items: { type: string }
        isModerated: { type: boolean }
        moderationStatus: { type: string, enum: [approved, pending, rejected] }
        createdAt: { type: string }
    PostInput:
      type: object
      required: [userId, type, mediaUrl]
      properties:
        userId: { type: integer }
        type: { type: string, enum: [video, image] }
        title: { type: string }
        description: { type: string }
        mediaUrl: { type: string }
        thumbnailUrl: { type: string }
        musicTitle: { type: string }
        musicArtist: { type: string }
        musicUrl: { type: string }
        duration: { type: integer }
        tags:
          type: array
          items: { type: string }
    LikeInput:
      type: object
      required: [userId, liked]
      properties:
        userId: { type: integer }
        liked: { type: boolean }
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked: { type: boolean }
        likeCount: { type: integer }

    # ── LIVE STREAM ──
...
    Comment:
      type: object
      required: [id, postId, userId, text, createdAt]
      properties:
        id: { type: integer }
        postId: { type: integer }
        userId: { type: integer }
        user: { $ref: "#/components/schemas/User" }
        text: { type: string }
        likeCount: { type: integer }
        isFlagged: { type: boolean }
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [userId, text]
      properties:
        userId: { type: integer }
        text: { type: string }

    # ── NOTIFICATIONS ──
...